Move last_modified function to utils module
