+++ src/sc/photogallery/browser/zip.py
@@ -2,7 +2,7 @@
 from plone import api
 from sc.photogallery import _
 from sc.photogallery.config import HAS_ZIPEXPORT
-from sc.photogallery.browser.view import last_modified
+from sc.photogallery.utils import last_modified
 from zope.component import getMultiAdapter
 from zope.publisher.browser import BrowserPage
 from ZPublisher.Iterators import filestream_iterator
